requirements.txt
streamlit==1.35.0
sentence-transformers==2.6.1
scikit-learn==1.4.2
numpy==1.26.4
scipy==1.11.4
matplotlib==3.7.5
deep-translator==1.11.4
python-dotenv==1.0.1
google-generativeai==0.8.5
PyMuPDF==1.24.2
